% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{avg}
\alias{avg}
\title{AVG distance (lowlevel function)}
\usage{
avg(P, Q, testNA)
}
\arguments{
\item{P}{a numeric vector storing the first distribution.}

\item{Q}{a numeric vector storing the second distribution.}

\item{testNA}{a logical value indicating whether or not distributions shall be checked for \code{NA} values.}
}
\description{
The lowlevel function for computing the avg distance.
}
\examples{
avg(P = 1:10/sum(1:10), Q = 20:29/sum(20:29), testNA = FALSE)
}
\author{
Hajk-Georg Drost
}
