% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binned.kernel.est.R
\name{binned.kernel.est}
\alias{binned.kernel.est}
\title{Kernel Density Estimation}
\usage{
binned.kernel.est(data, kernel = "normal", bandwidth = NULL,
  canonical = FALSE, scalest = "minim", level = 2L, gridsize = 401L,
  range.data = range(data), truncate = TRUE)
}
\arguments{
\item{data}{a numeric vector containing the sample on which the kernel density estimate is to be constructed.}

\item{kernel}{}

\item{bandwidth}{}

\item{canonical}{}

\item{scalest}{}

\item{level}{}

\item{gridsize}{}

\item{range.data}{}

\item{truncate}{}
}
\description{
This function implements an interface to the kernel density estimation functions provided in the \pkg{KernSmooth} package.
}
\author{
Hajk-Georg Drost
}

