% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dist.diversity.R
\name{dist.diversity}
\alias{dist.diversity}
\title{Distance Diversity between Probability Density Functions}
\usage{
dist.diversity(x, y, p = 2)
}
\arguments{
\item{x}{a numeric vector (probability density function).}

\item{y}{a numeric vector (probability density function).}

\item{p}{power of the Minkowski distance.}
}
\description{
This function computes all distance values between two probability density functions that are available in \code{\link{getDistMethods}}
and returns a vector storing the corresponding distance measures. This vector is \emph{named distance diversity vector}.
}
\examples{
dist.diversity(1:10/sum(1:10), 20:29/sum(20:29))
}
\author{
Hajk-Georg Drost
}

