% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{getDistMethods}
\alias{getDistMethods}
\title{Get method names for \code{distance}}
\usage{
getDistMethods()
}
\description{
This function returns the names of the methods that can be applied
to compute distances between probability density functions using the \code{\link{distance}}
function.
}
\examples{
getDistMethods()
}
\author{
Hajk-Georg Drost
}

