% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gJSD.R
\name{gJSD}
\alias{gJSD}
\title{Generalized Jensen-Shannon Divergence}
\usage{
gJSD(x)
}
\arguments{
\item{x}{a probability matrix.}
}
\value{
The Jensen-Shannon divergence between all possible combinations of comparisons.
}
\description{
This function computes the Generalized Jensen-Shannon Divergence of a probability matrix with equal weights.
}
\details{
Function to compute the Generalized Jensen-Shannon Divergence
}
\examples{
Prob <- Probability(PhyloExpressionSetExample)

# compute the Generalized JSD comparing the PS probability matrix
gJSD(Prob)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{KL}}, \code{\link{KL.Matrix}}, \code{\link{E}}, \code{\link{JSD}}
}

