% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jensen_difference}
\alias{jensen_difference}
\title{Jensen difference (lowlevel function)}
\usage{
jensen_difference(P, Q, testNA, unit)
}
\arguments{
\item{P}{a numeric vector storing the first distribution.}

\item{Q}{a numeric vector storing the second distribution.}

\item{testNA}{a logical value indicating whether or not distributions shall be checked for \code{NA} values.}

\item{unit}{type of \code{log} function. Option are 
\itemize{
\item \code{unit = "log"}
\item \code{unit = "log2"}
\item \code{unit = "log10"}   
}}
}
\description{
The lowlevel function for computing the jensen_difference distance.
}
\examples{
jensen_difference(P = 1:10/sum(1:10), Q = 20:29/sum(20:29), testNA = FALSE, unit = "log2")
}
\author{
Hajk-Georg Drost
}
