% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KL.Matrix.R
\name{KL.Matrix}
\alias{KL.Matrix}
\title{Kullback-Leibler Divergence of a Probability Matrix}
\usage{
KL.Matrix(x, test.na = TRUE, unit = "log2")
}
\arguments{
\item{x}{a probability matrix.}

\item{test.na}{a boolean value specifying whether input vectors shall be tested for NA values.}

\item{unit}{a character string specifying the logarithm unit that shall be used to compute distances that depend on log computations.}
}
\value{
a matrix of pairwise KL divergences between all combinations of possible comparisons.
}
\description{
This function computes the cross entropy (Kullback-Leibler Divergence) of
all combinations of stage specific probability distributions.
}
\examples{
Prob <- cbind(1:10/sum(1:10), 20:29/sum(20:29), 30:39/sum(30:39))

# compute the KL matrix of a given probability matrix
KLMatrix <- KL.Matrix(Prob)

# plot a heatmap of the corresponding KL matrix
heatmap(KLMatrix)
}
\author{
Hajk-Georg Drost
}
\references{
Cover Thomas M. and Thomas Joy A. 2006. "Elements of Information
Theory". \emph{John Wiley & Sons}.
}
\seealso{
\code{\link{gJSD}}, \code{\link{KL}}, \code{\link{H}}, \code{\link{CE}}, \code{\link{JE}}
}

