% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kullback_leibler_distance}
\alias{kullback_leibler_distance}
\title{kullback-Leibler distance (lowlevel function)}
\usage{
kullback_leibler_distance(P, Q, testNA, unit)
}
\arguments{
\item{P}{a numeric vector storing the first distribution.}

\item{Q}{a numeric vector storing the second distribution.}

\item{testNA}{a logical value indicating whether or not distributions shall be checked for \code{NA} values.}

\item{unit}{type of \code{log} function. Option are 
\itemize{
\item \code{unit = "log"}
\item \code{unit = "log2"}
\item \code{unit = "log10"}   
}}
}
\description{
The lowlevel function for computing the kullback_leibler_distance distance.
}
\examples{
kullback_leibler_distance(P = 1:10/sum(1:10), Q = 20:29/sum(20:29), testNA = FALSE, unit = "log2")
}
\author{
Hajk-Georg Drost
}
