% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lin.cor.R
\name{lin.corr}
\alias{lin.corr}
\title{Linear Correlation}
\usage{
lin.corr(x, y = NULL, method = "pearson")
}
\arguments{
\item{x}{a numeric \code{vector}, \code{matrix}, or \code{data.frame}.}

\item{y}{a numeric \code{vector} that should be correlated with \code{x}.}
}
\description{
This function computed the linear correlation between two vectors or a correlation matrix for an input matrix.

The following methods to compute linear correlations are implemented in this function:
}
\details{
\itemize{
  \item \code{method = "pearson"} : Pearson's correlation coefficient (centred).
  \item \code{method = "pearson2"} : Pearson's uncentred correlation coefficient.
  \item \code{method = "sq_pearson"} . Squared Pearson's correlation coefficient.
  \item \code{method = "kendall"} : Kendall's correlation coefficient.
  \item \code{method = "spearman"} : Spearman's correlation coefficient.
  }

 Further Details:

 \itemize{
 \item \emph{Pearson's correlation coefficient (centred)} :
 }
}
\author{
Hajk-Georg Drost
}

