% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minkowski}
\alias{minkowski}
\title{Minkowski distance (lowlevel function)}
\usage{
minkowski(P, Q, n, testNA)
}
\arguments{
\item{P}{a numeric vector storing the first distribution.}

\item{Q}{a numeric vector storing the second distribution.}

\item{n}{index for the minkowski exponent.}

\item{testNA}{a logical value indicating whether or not distributions shall be checked for \code{NA} values.}
}
\description{
The lowlevel function for computing the minkowski distance.
}
\examples{
minkowski(P = 1:10/sum(1:10), Q = 20:29/sum(20:29), n = 2, testNA = FALSE)
}
\author{
Hajk-Georg Drost
}
