% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Probability.R
\name{Probability}
\alias{Probability}
\title{Generate a Probability Matrix}
\usage{
Probability(ExpressionSet)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or
DivergenceExpressionSet object.}
}
\value{
a numeric matrix storing the probability distributions of PS or DS
for each developmental stage s.
}
\description{
This function computes the probability of each
Phylostratum/Divergence Stratum for each developmental stage \code{s}. The
probability of a given phylostratum \code{ps} is defined as the sum of expression
levels in stage \code{s} of genes corresponding to \code{ps} divided by the overall sum of
expression levels in stage \code{s}.
}
\details{
\deqn{P_s(PS=ps) = \sum ( e_is * \delta(ps_i = ps) ) / \sum e_is}

Where \eqn{e_is} denotes the expression level of gene \eqn{i} in stage
\bold{\eqn{s}}, \bold{\eqn{\delta}} denotes the \bold{Kronecker delta}, and
\bold{\eqn{ps_i}} denotes the corresponding phylostratum of gene \eqn{i}.

The goal of this function is to model the discrete PS or DS values as probabilities.
}
\examples{
# example data sets
data(PhyloExpressionSetExample)
data(DivergenceExpressionSetExample)

# generate the Phylostratum Probability Matrix for the given PhyloExpressionSet
Probability(PhyloExpressionSetExample)

# generate the Divergence Stratum Probability Matrix for the given DivergenceExpressionSet
Probability(DivergenceExpressionSetExample)
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{Phylotranscriptomics_Probability}},
\code{\link{Phylotranscriptomics_JointProbability}},
\code{\link{JointProbability}}, \code{\link{Phylotranscriptomics_Entropy}},
\code{\link{Entropy}}
}

