% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Sampled.Entropy.Matrix.R
\name{Sampled.Entropy.Matrix}
\alias{Sampled.Entropy.Matrix}
\title{Perform Sampling for Entropy}
\usage{
Sampled.Entropy.Matrix(ExpressionSet, permutations = 1000, parallel = FALSE)
}
\arguments{
\item{ExpressionSet}{a standard PhyloExpressionSet or
DivergenceExpressionSet object.}

\item{permutations}{a numeric value specifying the number of permutations
that shall be performed. Default is \code{permutations} = \code{1000}.}

\item{parallel}{a boolean value specifying whether multicore (parallel)
processing shall be performed on a multicore machine.}
}
\value{
a numeric \code{N x S} matrix containing Entropy
values for \code{S} developmental stages after \code{N} permutations.
}
\description{
This function receives a standard PhyloExpressionSet or
DivergenceExpressionSet, samples the Phylostratum or Divergence Stratum column of the corresponding
PhyloExpressionSet or DivergenceExpressionSet and furthermore, computes the
Entropy profile of this sampled PhyloExpressionSet or
DivergenceExpressionSet.

The resulting sample matrix consists of \code{N x S} elements, where S denotes the
total number of developmental stages stored in the corresponding
PhyloExpressionSet or DivergenceExpressionSet and \code{N} the number of permutations. This \code{N x S} sample matrix can
be used to perform further test statistics (sampling test) to quantify the
statistical significance of any observed Entropy pattern.
}
\examples{
# read example phylotranscriptomics data
data(PhyloExpressionSetExample)
data(DivergenceExpressionSetExample)

# perform N = 1000 sampling computations sequencially
sMatrix <- Sampled.Entropy.Matrix( ExpressionSet = PhyloExpressionSetExample,
                                   permutations  = 1000 )

\dontrun{

# perform N = 1000 sampling computations on a multicore machine to speed up computations
psMatrix <- Sampled.Entropy.Matrix( ExpressionSet = PhyloExpressionSetExample,
                                    permutations  = 1000,
                                    parallel      = TRUE )
}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{Entropy}}
}

