% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squared_chord}
\alias{squared_chord}
\title{Squared chord distance (lowlevel function)}
\usage{
squared_chord(P, Q, testNA)
}
\arguments{
\item{P}{a numeric vector storing the first distribution.}

\item{Q}{a numeric vector storing the second distribution.}

\item{testNA}{a logical value indicating whether or not distributions shall be checked for \code{NA} values.}
}
\description{
The lowlevel function for computing the squared_chord distance.
}
\examples{
squared_chord(P = 1:10/sum(1:10), Q = 20:29/sum(20:29), testNA = FALSE)
}
\author{
Hajk-Georg Drost
}
