% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{valid.distr}
\alias{valid.distr}
\title{Check the validity of the input probability distribution}
\usage{
valid.distr(x, test.na = TRUE)
}
\description{
Check the validity of the input probability distribution
}

