% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binned.kernel.est.R
\name{binned.kernel.est}
\alias{binned.kernel.est}
\title{Kernel Density Estimation}
\usage{
binned.kernel.est(data, kernel = "normal", bandwidth = NULL,
  canonical = FALSE, scalest = "minim", level = 2L, gridsize = 401L,
  range.data = range(data), truncate = TRUE)
}
\arguments{
\item{data}{a numeric vector containing the sample on which the kernel density estimate is to be constructed.}

\item{kernel}{}

\item{bandwidth}{}

\item{canonical}{}

\item{scalest}{}

\item{level}{}

\item{gridsize}{}

\item{range.data}{}

\item{truncate}{}
}
\description{
This function implements an interface to the kernel density estimation functions provided by the \pkg{KernSmooth} package.
}
\author{
Hajk-Georg Drost
}
\references{
Matt Wand (2015). KernSmooth: Functions for Kernel Smoothing Supporting Wand & Jones (1995). R package version 2.23-14. \url{http://CRAN.R-project.org/package=KernSmooth}

Henry Deng and Hadley Wickham (2011). Density estimation in R. \url{http://vita.had.co.nz/papers/density-estimation.pdf}.
}

