% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance between Probability Density Functions}
\usage{
distance(x, y, method = "euclidean", p = NULL)
}
\arguments{
\item{x}{a numeric vector (probability density function).}

\item{y}{a numeric vector (probability density function).}

\item{method}{a character string specifying the distance measure that shall be computed.}

\item{p}{power of the Minkowski distance.}
}
\description{
This functions computes the distance/dissimilarity between two probability density functions.
}
\details{
The following distance measures are implemented in this function:

\itemize{
\item L_p Minkowski family
\itemize{
\item Euclidean : \eqn{d = sqrt( \sum | P_i - Q_i |^2)}
\item Manhattan : \eqn{d = \sum | P_i - Q_i |}
\item Minkowski : \eqn{d = ( \sum | P_i - Q_i |^p)^1/p}
\item Chebyshev : \eqn{d = max | P_i - Q_i |}
}

\item L_1 family
\itemize{
\item Sorensen : \eqn{d = \sum | P_i - Q_i | / \sum (P_i + Q_i)}
\item Gower : \eqn{d = 1/d * \sum | P_i - Q_i |}
\item Soergel : \eqn{d = \sum | P_i - Q_i | / \sum max(P_i , Q_i)}
\item Kulczynski d : \eqn{d = \sum | P_i - Q_i | / \sum min(P_i , Q_i)}
\item Canberra : \eqn{d = \sum | P_i - Q_i | / (P_i + Q_i)}
\item Lorentzian : \eqn{d = \sum ln(1 + | P_i - Q_i |)}
}


\item Intersection family
\itemize{
\item Intersection : \eqn{s = \sum min(P_i , Q_i)}
\item Non-Intersection : \eqn{d = 1 - \sum min(P_i , Q_i)}
\item Wave Hedges : \eqn{d = \sum | P_i - Q_i | / max(P_i , Q_i)}
\item Czekanowski : \eqn{d = \sum | P_i - Q_i | / \sum | P_i + Q_i |}
\item Motyka : \eqn{d = \sum min(P_i , Q_i) / (P_i + Q_i)}
\item Kulczynski s : \eqn{d = 1 / \sum | P_i - Q_i | / \sum min(P_i , Q_i)}
\item Tanimoto : \eqn{d = \sum (max(P_i , Q_i) - min(P_i , Q_i)) / \sum max(P_i , Q_i)} ; equivalent to Soergel
\item Ruzicka : \eqn{s = \sum min(P_i , Q_i) / \sum max(P_i , Q_i)} ; equivalent to 1 - Tanimoto = 1 - Soergel
}

\item Inner Product family
\itemize{
\item Inner Product : \eqn{s = \sum P_i * Q_i}
\item Harmonic mean : \eqn{s = 2 * \sum (P_i * Q_i) / (P_i + Q_i)}
\item Cosine : \eqn{s = \sum (P_i * Q_i) / sqrt(\sum P_i^2) * sqrt(\sum Q_i^2)}
\item Kumar-Hassebrook (PCE) : \eqn{s = \sum (P_i * Q_i) / (\sum P_i^2 + \sum Q_i^2 - \sum (P_i * Q_i))}
\item Jaccard : \eqn{d = 1 - \sum (P_i * Q_i) / (\sum P_i^2 + \sum Q_i^2 - \sum (P_i * Q_i))} ; equivalent to 1 - Kumar-Hassebrook
\item Dice : \eqn{d = \sum (P_i - Q_i)^2 / (\sum P_i^2 + \sum Q_i^2)}
}

\item Squared-chord family
\itemize{
\item Fidelity : \eqn{s = \sum sqrt(P_i * Q_i)}
\item Bhattacharyya : \eqn{d = - ln \sum sqrt(P_i * Q_i)}
\item Hellinger : \eqn{d = 2 * sqrt( 1 - \sum sqrt(P_i * Q_i))}
\item Matusita : \eqn{d = sqrt( 2 - 2 * \sum sqrt(P_i * Q_i))}
\item Squared-chord : \eqn{d = \sum ( sqrt(P_i) - sqrt(Q_i) )^2}
}


\item Squared L_2 family (\eqn{\Chi}^2 squared family)
\itemize{
\item Squared Euclidean : \eqn{d = \sum ( P_i - Q_i )^2}
\item Pearson \eqn{\Chi}^2 : \eqn{d = \sum ( (P_i - Q_i )^2 / Q_i )}
\item Neyman \eqn{\Chi}^2 : \eqn{d = \sum ( (P_i - Q_i )^2 / P_i )}
\item Squared \eqn{\Chi}^2 : \eqn{d = \sum ( (P_i - Q_i )^2 / (P_i + Q_i) )}
\item Probabilistic Symmetric \eqn{\Chi}^2 : \eqn{d = 2 *  \sum ( (P_i - Q_i )^2 / (P_i + Q_i) )}
\item Divergence : \eqn{\Chi}^2 : \eqn{d = 2 *  \sum ( (P_i - Q_i )^2 / (P_i + Q_i)^2 )}
\item Clark : \eqn{d = sqrt ( \sum (| P_i - Q_i | / (P_i + Q_i))^2 )}
\item Additive Symmetric \eqn{\Chi}^2 : \eqn{d = \sum ( ((P_i - Q_i)^2 * (P_i + Q_i)) / (P_i * Q_i) ) }
}

\item Shannon's entropy family
\itemize{
\item Kullback-Leibler : \eqn{d = \sum P_i * log(P_i / Q_i)}
\item Jeffreys : \eqn{d = \sum (P_i - Q_i) * log(P_i / Q_i)}
\item K divergence : \eqn{d = \sum P_i * log(2 * P_i / P_i + Q_i)}
\item Topsoe : \eqn{d = \sum ( P_i * log(2 * P_i / P_i + Q_i) ) + ( Q_i * log(2 * Q_i / P_i + Q_i) )}
\item Jensen-Shannon :  \eqn{d = 0.5 * ( \sum P_i * log(2 * P_i / P_i + Q_i) + \sum Q_i * log(2 * Q_i / P_i + Q_i))}
\item Jensen difference : \eqn{d = \sum ( (P_i * log(P_i) + Q_i * log(Q_i) / 2) - (P_i + Q_i / 2) * log(P_i + Q_i / 2) )}
}

\item Combinations
\itemize{
\item Taneja : \eqn{d = \sum ( P_i + Q_i / 2) * log( P_i + Q_i / ( 2 * sqrt( P_i * Q_i)) )}
\item Kumar-Johnson : \eqn{d = \sum (P_i^2 - Q_i^2)^2 / 2 * (P_i * Q_i)^1.5}
\item Avg(L_1, L_n) : \eqn{d = \sum | P_i - Q_i| + max{ | P_i - Q_i |} / 2}
}

}
}
\examples{
distance(1:10/sum(1:10), 20:29/sum(20:29), method = "euclidean")
}
\author{
Hajk-Georg Drost
}

