% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance between Probability Density Functions}
\usage{
distance(x, y, method = "euclidean", n = NULL)
}
\arguments{
\item{x}{a numeric vector (probability density function).}

\item{y}{a numeric vector (probability density function).}

\item{method}{a character string specifying the distance measure that shall be computed.}
}
\description{
This functions computes the distance/dissimilarity between two probability density functions.
}
\details{
The following distance measures are implemented in this function:
}
\examples{
distance(1:10, 20:29, method = "euclidean")
}
\author{
Hajk-Georg Drost
}

