% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance between Probability Density Functions}
\usage{
distance(x, y, method = "euclidean", p = NULL)
}
\arguments{
\item{x}{a numeric vector (probability density function).}

\item{y}{a numeric vector (probability density function).}

\item{method}{a character string specifying the distance measure that shall be computed.}
}
\description{
This functions computes the distance/dissimilarity between two probability density functions.
}
\details{
The following distance measures are implemented in this function:

\itemize{
\item L_p Minkowski family
\itemize{
\item Euclidean : \eqn{d = sqrt( \sum | P_i - Q_i |^2)}
\item Manhatten : \eqn{d = \sum | P_i - Q_i |}
\item Minkowski : \eqn{d = ( \sum | P_i - Q_i |^p)^1/p}
\item Chebyshev : \eqn{d = max | P_i - Q_i |}
}

\item L_1 family
\itemize{
\item Sorensen : \eqn{d = \sum | P_i - Q_i | / \sum (P_i + Q_i)}
\item Gower : \eqn{d = 1/d * \sum | P_i - Q_i |}
\item Soergel : \eqn{d = \sum | P_i - Q_i | / \sum max(P_i , Q_i)}
\item Kulczynski d : \eqn{d = \sum | P_i - Q_i | / \sum min(P_i , Q_i)}
\item Canberra : \eqn{d = \sum | P_i - Q_i | / (P_i + Q_i)}
\item Lorentzian : \eqn{d = \sum ln(1 + | P_i - Q_i |)}
}


\item Intersection family
\itemize{
\item Intersection : \eqn{d = \sum min(P_i , Q_i)}
\item Non-Intersection : \eqn{d = 1 - \sum min(P_i , Q_i)}
\item Wave Hedges : \eqn{d = \sum | P_i - Q_i | / max(P_i , Q_i)}
\item Czekanowski : \eqn{d = \sum | P_i - Q_i | / \sum | P_i + Q_i |}
\item Motyka : \eqn{d = \sum min(P_i , Q_i) / (P_i + Q_i)}
\item Kulczynski s : \eqn{d = 1.0 / \sum | P_i - Q_i | / \sum min(P_i , Q_i)}
\item Tanimoto : \eqn{d = \sum (max(P_i , Q_i) - min(P_i , Q_i)) / \sum max(P_i , Q_i)} ; equivalent to Soergel
\item Ruzicka : \eqn{d = \sum min(P_i , Q_i) / \sum max(P_i , Q_i)} ; equivalent to 1 - Tanimoto = 1 - Soergel
}

\item Inner Product family
\itemize{
\item Inner Product : \eqn{d = \sum P_i * Q_i}
\item Harmonic mean : \eqn{d = 2 * \sum (P_i * Q_i) / (P_i + Q_i)}
\item Cosine : \eqn{d = \sum (P_i * Q_i) / sqrt(\sum P_i^2) * sqrt(\sum Q_i^2)}
\item Kumar-Hassebrook (PCE) : \eqn{d = \sum (P_i * Q_i) / (\sum P_i^2 + \sum Q_i^2 - \sum (P_i * Q_i))}
\item Jaccard : \eqn{d = 1 - \sum (P_i * Q_i) / (\sum P_i^2 + \sum Q_i^2 - \sum (P_i * Q_i))} ; equivalent to 1 - Kumar-Hassebrook
\item Dice : \eqn{d = \sum (P_i - Q_i)^2 / (\sum P_i^2 + \sum Q_i^2)}
}
\item Shannon's entropy family
\itemize{
\item Kullback-Leibler : \eqn{KL(P || Q) = \sum P(P) * log2(P(P) / P(Q)) = H(P,Q) - H(P)}
\item Jensen-Shannon : \eqn{JSP(P || Q) = 0.5 * (KL(P || R) + KL(Q || R))}, where \eqn{R = 0.5 * (P + Q)} denotes the mid-point of the probability
vectors P and Q
}
}
}
\examples{
distance(1:10/sum(1:10), 20:29/sum(20:29), method = "euclidean")
}
\author{
Hajk-Georg Drost
}

