% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/H.R
\name{H}
\alias{H}
\title{Shannon's Entropy \eqn{H(X)}}
\usage{
H(x)
}
\arguments{
\item{x}{a numeric probability vector \eqn{P(X)} for which
Shannon's Entropy \eqn{H(X)} shall be computed.}
}
\value{
a numeric value representing Shannon's Entropy in bit.
}
\description{
Compute the Shannon's Entropy \eqn{H(X) = - \sum P(X) * log2(P(X))} based on a
given probability vector \eqn{P(X)}.
}
\details{
This function might be useful to fastly compute Shannon's Entropy for any
given probability vector.
}
\examples{
H(1:10/sum(1:10))
}
\author{
Hajk-Georg Drost
}
\references{
Shannon, Claude E. 1948. "A Mathematical Theory of
Communication". \emph{Bell System Technical Journal} \bold{27} (3): 379-423.
}
\seealso{
\code{\link{JE}}, \code{\link{CE}}, \code{\link{KL}}, \code{\link{JSD}}, \code{\link{gJSD}}
}

