% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/KL.Matrix.R
\name{KL.Matrix}
\alias{KL.Matrix}
\title{Kullback-Leibler Divergence of a Probability Matrix}
\usage{
KL.Matrix(x)
}
\arguments{
\item{x}{a probability matrix.}
}
\value{
a matrix of pairwise KL divergences between all combinations of possible comparisons.
}
\description{
This function computes the cross entropy (Kullback-Leibler Divergence) of
all combinations of stage specific probability distributions.
}
\examples{
# read standard phylotranscriptomics data
data(PhyloExpressionSetExample)

# compute the KL matrix of a given PhyoExpressionSet
KLMatrix <- KL.Matrix(Probability(PhyloExpressionSetExample))

# plot a heatmap of the corresponding KL matrix
heatmap(KLMatrix)
}
\author{
Hajk-Georg Drost
}
\references{
Cover Thomas M. and Thomas Joy A. 2006. "Elements of Information
Theory". \emph{John Wiley & Sons}.
}
\seealso{
\code{\link{gJSD}}, \code{\link{KL}}, \code{\link{Probability}}
}

